Removed sbuxton open and close functions
Instead, sbuxton_register_notify opens a client if one is not open,
then upon success of the registration it increments an int indicating
the number of registered notifications. If the number is not 0, then
other functions will not close the connection at the end of a call.
Unregister_notify will decrement the number upon success of a call.